refactor: move energy out of infrastructure
