+++ src/libecalc/domain/energy/__init__.py
