+++ src/libecalc/domain/energy/component_energy_context.py
