+++ src/libecalc/domain/infrastructure/energy_components/asset/asset.py
@@ -1,6 +1,6 @@
 from libecalc.common.component_type import ComponentType
 from libecalc.common.string.string_utils import generate_id
-from libecalc.domain.infrastructure.energy import EnergyComponent
+from libecalc.domain.energy import EnergyComponent
 from libecalc.domain.infrastructure.energy_components.installation.installation import Installation
 from libecalc.dto.component_graph import ComponentGraph
