+++ src/libecalc/domain/infrastructure/energy_components/compressor/component_dto.py
@@ -4,7 +4,7 @@ from libecalc.common.component_type import ComponentType
 from libecalc.common.consumption_type import ConsumptionType
 from libecalc.common.string.string_utils import generate_id
 from libecalc.common.time_utils import Period
-from libecalc.domain.infrastructure.energy import EnergyComponent
+from libecalc.domain.energy import EnergyComponent
 from libecalc.domain.infrastructure.energy_components.component_validation_error import (
     ComponentValidationException,
     ModelValidationError,
