+++ src/libecalc/dto/component_graph.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from libecalc.common.component_info.component_level import ComponentLevel
 from libecalc.common.component_type import ComponentType
 from libecalc.common.graph import Graph, NodeID
-from libecalc.domain.infrastructure.energy import EnergyComponent
+from libecalc.domain.energy import EnergyComponent
 from libecalc.dto.node_info import NodeInfo
